Remove splice of `fromInteger` in TH
They were useless and were triggering a default type selection and associated warnings.

+++ src/PyF/Internal/QQ.hs
@@ -116,11 +116,11 @@ toFormat _ (Replacement expr y) = do
 
 changePrec :: Precision -> Q Exp
 changePrec PrecisionDefault = [| Just 6 |]
-changePrec (Precision n) = [| Just (fromIntegral n) |]
+changePrec (Precision n) = [| Just n |]
 
 changePrec' :: Precision ->  Q Exp
 changePrec' PrecisionDefault = [| Nothing |]
-changePrec' (Precision n) = [| Just (fromIntegral n) |]
+changePrec' (Precision n) = [| Just n |]
 
 toGrp :: Maybe Char -> Int -> Q Exp
 toGrp mb a = [| grp |]
